+++ sentry-rails/CHANGELOG.md
@@ -6,6 +6,7 @@
   - Fixes [#1071](https://github.com/getsentry/sentry-ruby/issues/1071)
 - Make sentry-rails a Rails engine and provide default job class for async [#1181](https://github.com/getsentry/sentry-ruby/pull/1181)
 - Allow users to configure ActiveJob adapters to ignore [#1256](https://github.com/getsentry/sentry-ruby/pull/1256)
+- Tag `job_id` and `provider_job_id` on ActiveJob events [#1259](https://github.com/getsentry/sentry-ruby/pull/1259)
 
 ## Unreleased